requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asteroid-filterbanks==0.4.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
datasets==2.14.6
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.7
docopt==0.6.2
einops==0.7.0
emoji==2.8.0
executing==2.0.0
facenet-pytorch==2.5.3
fastjsonschema==2.18.1
ffmpeg-python==0.2.0
ffmpegio==0.8.4
ffmpegio-core==0.8.4
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
greenlet==3.0.1
h5py==3.10.0
huggingface-hub==0.17.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
julius==0.2.7
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
librosa==0.10.1
-e git+https://github.com/cvg/LightGlue.git@29f3e449efa1994758b8a16299d2816028dca65b#egg=lightglue
lightning==2.1.1
lightning-utilities==0.9.0
llvmlite==0.41.1
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
omegaconf==2.3.0
onnxruntime-gpu==1.16.2
opencv-python==4.8.1.78
optuna==3.4.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.11.0
pluggy==1.3.0
pooch==1.8.0
primePy==1.3
proglog==0.1.10
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyannote.audio==3.0.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==14.0.1
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.1.1
pytorch-metric-learning==2.3.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.12.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
semver==3.0.2
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==0.5.15
SQLAlchemy==2.0.23
stack-data==0.6.3
stanza==1.6.1
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
terminado==0.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.14.1
torch==2.1.0+cu118
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.1.0+cu118
torchdata==0.7.0
torchmetrics==1.2.0
torchtext==0.16.0+cpu
torchvision==0.16.0+cu118
tornado==6.3.3
tqdm==4.64.1
traitlets==5.11.2
transformers==4.35.0
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.208
uri-template==1.3.0
urllib3==1.26.13
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.2
filterpy==1.4.5
